# docs/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "MentisEterna API",
        "version": "1.0.0",
        "description": "Personal note-taking app with vector search, note-type plugins, and file attachments. All API endpoints except `/login` and `/webauthn/*` require a Bearer token (obtained via password login or passkey login)."
    },
    "servers": [
        { "url": "http://localhost:8080", "description": "Local development" }
    ],
    "security": [{ "bearerAuth": [] }, { "cookieAuth": [] }],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "description": "Session token obtained from POST /login or passkey login."
            },
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "auth_token",
                "description": "Cookie set automatically after login. Used by browser-embedded resource requests (img, video, etc.)."
            }
        },
        "schemas": {
...
            "UpdateNoteRequest": {
                "allOf": [{ "$ref": "#/components/schemas/CreateNoteRequest" }]
            },
            "LoginRequest": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                    "username": { "type": "string" },
                    "password": { "type": "string" }
                }
            },
            "LoginResponse": {
                "type": "object",
                "required": ["token", "expires_at"],
                "properties": {
                    "token": {
                        "type": "string",
                        "description": "Bearer token for subsequent requests"
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "Token expiry timestamp"
                    }
                }
            },
...
        }
    },
    "paths": {
...
        "/login": {
            "post": {
                "summary": "Password login",
                "operationId": "login",
                "tags": ["Auth"],
                "security": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful. The response includes a Bearer token and sets an auth_token cookie.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing username/password or invalid JSON"
                    },
                    "401": { "description": "Invalid credentials" }
                }
            }
        },
        "/webauthn/register/begin": {
            "get": {
                "summary": "Begin passkey registration",
                "operationId": "webauthnRegisterBegin",
                "tags": ["Auth"],
                "description": "Requires an existing session. Returns WebAuthn creation options.",
                "responses": {
                    "200": { "description": "Creation options JSON" },
                    "401": { "description": "Not authenticated" }
                }
            }
        },
        "/webauthn/register/finish": {
            "post": {
                "summary": "Finish passkey registration",
                "operationId": "webauthnRegisterFinish",
                "tags": ["Auth"],
                "description": "Completes WebAuthn registration with the attestation response from the browser.",
                "responses": {
                    "201": { "description": "Passkey registered successfully" },
                    "400": { "description": "Invalid attestation" }
                }
            }
        },
...
        "/webauthn/login/finish": {
            "post": {
                "summary": "Finish passkey login",
                "operationId": "webauthnLoginFinish",
                "tags": ["Auth"],
                "security": [],
                "description": "Completes WebAuthn login with the assertion response from the browser. Returns a session token.",
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "400": { "description": "Invalid assertion" }
                }
            }
        },
...
        "/notes": {
            "get": {
                "summary": "List all notes",
                "operationId": "listNotes",
                "tags": ["Notes"],
                "description": "Returns a lightweight summary of all notes, ordered by pinned status then last-updated time descending. Plugin data is not included in this response.",
                "responses": {
                    "200": {
                        "description": "List of note summaries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NoteSummary"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a note",
                "operationId": "createNote",
                "tags": ["Notes"],
                "description": "Creates a new note. If title is empty/blank, it defaults to 'Untitled'. If type is empty, it defaults to 'standard'. Returns the full note detail with plugin config/view populated.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateNoteRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Note created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteDetail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or plugin validation failed"
                    }
                }
            }
        },
        "/notes/search": {
            "get": {
                "summary": "Semantic search",
                "operationId": "searchNotes",
                "tags": ["Notes"],
                "description": "Performs a semantic (vector) search across note bodies, OCR text, and STT text. Results are ranked by cosine distance. Falls back to empty results if VSS is unavailable.",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "description": "Search query string",
                        "schema": { "type": "string" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results sorted by relevance",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SearchResult"
                                    }
                                }
                            }
                        }
                    },
                    "400": { "description": "Missing or empty query parameter" }
                }
            }
        },
        "/notes/{id}": {
            "get": {
                "summary": "Get a note",
                "operationId": "getNote",
                "tags": ["Notes"],
                "description": "Returns the full note detail including plugin config, plugin view, tags, and attachments.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteDetail"
                                }
                            }
                        }
                    },
                    "404": { "description": "Note not found" }
                }
            },
            "put": {
                "summary": "Update a note",
                "operationId": "updateNote",
                "tags": ["Notes"],
                "description": "Updates a note's title, body, parent, type, custom_data, and tags. Creates a new history entry. Returns the full note detail.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateNoteRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteDetail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid JSON or plugin validation failed"
                    },
                    "404": { "description": "Note not found" }
                }
            },
            "delete": {
                "summary": "Delete a note",
                "operationId": "deleteNote",
                "tags": ["Notes"],
                "description": "Deletes a note and all its history, plugin data (via CASCADE), and VSS embedding. Attached files with no remaining references are soft-deleted.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "responses": {
                    "204": { "description": "Note deleted" },
                    "404": { "description": "Note not found" }
                }
            }
        },
        "/notes/{id}/history": {
            "get": {
                "summary": "Get note edit history",
                "operationId": "getNoteHistory",
                "tags": ["Notes"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of history entries (newest first)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NoteUpdate"
                                    }
                                }
                            }
                        }
                    },
                    "404": { "description": "Note not found" }
                }
            }
        },
        "/notes/{id}/children": {
            "get": {
                "summary": "Get child notes",
                "operationId": "getNoteChildren",
                "tags": ["Notes"],
                "description": "Returns all direct children of this note, ordered by creation date ascending. Each child includes a child_count.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of child notes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ChildNote"
                                    }
                                }
                            }
                        }
                    },
                    "404": { "description": "Parent note not found" }
                }
            }
        },
        "/notes/{id}/ancestors": {
            "get": {
                "summary": "Get ancestor breadcrumb chain",
                "operationId": "getNoteAncestors",
                "tags": ["Notes"],
                "description": "Returns the full ancestor chain from root to this note (root-first order).",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of ancestor notes (root → self)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NoteSummary"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/notes/{id}/pin": {
            "post": {
                "summary": "Toggle pin status",
                "operationId": "setNotePin",
                "tags": ["Notes"],
                "description": "Sets the pinned status of a note. Returns the full note detail.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PinRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pin status updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteDetail"
                                }
                            }
                        }
                    },
                    "404": { "description": "Note not found" }
                }
            }
        },
...
        "/notes/{id}/files": {
            "post": {
                "summary": "Upload file attachment",
                "operationId": "uploadAttachment",
                "tags": ["Files"],
                "description": "Uploads a file as a note attachment. The file is persisted to S3 (or local cache) and an attachment reference is created. OCR/STT is enqueued for images/audio.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" },
                        "description": "Note ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "required": ["file"],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "The file to upload"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "File uploaded",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AttachmentUploadResponse"
                                }
                            }
                        }
                    },
                    "404": { "description": "Note not found" },
                    "503": { "description": "Media service not configured" }
                }
            }
        },
...
        "/notes/{id}/files/{fileID}": {
            "delete": {
                "summary": "Delete file attachment",
                "operationId": "deleteAttachment",
                "tags": ["Files"],
                "description": "Removes the attachment reference between the note and file. If no other references exist, the file is soft-deleted.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" },
                        "description": "Note ID"
                    },
                    {
                        "name": "fileID",
                        "in": "path",
                        "required": true,
                        "schema": { "type": "integer", "format": "int64" },
                        "description": "File ID"
                    }
                ],
                "responses": {
                    "204": { "description": "Attachment reference removed" },
                    "400": { "description": "Invalid path" },
                    "503": { "description": "Media service not configured" }
                }
            }
        },
...
    }
}